+++ freight_forwarder/container/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8; -*-
 # flake8: noqa
-__author__ = 'alexander'
 from .container import Container